Remove streaming module and restore buffered edge coverage
