+++ src/session/mod.rs
@@ -173,12 +173,13 @@ pub mod handlers;
 pub mod metrics_ext;
 pub mod mode_state;
 pub mod precheck;
+pub(crate) mod response_buffer;
 pub(crate) mod retry;
 pub(crate) mod routing;
 pub mod session_error;
 pub(crate) mod shared_client_writer;
 pub mod state;
-pub mod streaming;
+pub mod tail_buffer;
 
 pub use auth_state::AuthState;
 pub use backend::format_hex_preview;
