+++ src/session/tail_buffer.rs
